Update and delete controller are modified to handle invalid note id

+++ src/controllers/notesControllers.js
@@ -1,4 +1,5 @@
 import Note from "../models/Note.js"
+import mongoose from "mongoose";
 
 export const getAllNotes = async (req, res) => {
     try {
@@ -25,7 +26,12 @@ export const createNewNote = async (req, res) => {
 
 export const updateNote = async (req, res) => {
     try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: "Invalid note ID" });
+        }
+
         const { title, content } = req.body;
+
         const updatedNote = await Note.findByIdAndUpdate(req.params.id,
             { title, content },
             {
@@ -33,6 +39,11 @@ export const updateNote = async (req, res) => {
                 runValidators: true,  // validates schema
             }
         )
+
+        if (!updatedNote) {
+            return res.status(404).json({ message: "Note not found" })
+        }
+
         res.status(200).json(updatedNote)
     } catch (error) {
         console.error("Error in updateNote controller", error)
@@ -42,10 +53,20 @@ export const updateNote = async (req, res) => {
 
 export const deleteNote = async (req, res) => {
     try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: "Invalid note ID" });
+        }
+
         const deleteNote = await Note.findByIdAndDelete(req.params.id)
-        res.status(200).json(deleteNote)
+
+        if (!deleteNote) {
+            return res.status(404).json({ message: "Note not found" })
+        }
+
+        res.status(200).json({ message: "Note deleted successfully!!" })
     } catch (error) {
         console.error("Error in deleteNote controller", error)
-        res.status(500).json({ messgae: "Internal Server Error"})
+
+        res.status(500).json({ messgae: "Internal Server Error" })
     }
 }
